Commit the first draft of a ding2tei converter
It doesn't "work" yet. The script actually only tokenizes the format and
provides a list of elements and events and some unparsed tokens. The tokenizing
is not complete yet, the elements and their typing is still incromprehensible
for others and documentation is missing, of course.
In the file READMe you find some pointers to the Ding format and the dictionary
source and ToDo is a cryptic description of things to do.